requirements.txt
absl-py==1.0.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
altair==4.2.2
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Automat==22.10.0
autopep8==1.7.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
blis==0.7.8
Boruta==0.3
cachetools==5.0.0
camelot-py==0.10.1
catalogue==1.0.2
certifi==2021.10.8
cffi==1.15.0
chardet==5.1.0
charset-normalizer==2.0.10
ci-info==0.3.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
colorlover==0.3.0
configobj==5.0.7
configparser==5.3.0
constantly==15.1.0
cryptography==38.0.3
cssselect==1.2.0
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.14
databricks-cli==0.17.4
dataframe-image==0.1.5
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
docker==6.0.1
docx==0.2.4
docx2txt==0.8
easydict==1.10
entrypoints==0.3
et-xmlfile==1.1.0
etelemetry==0.3.0
executing==0.8.2
fastjsonschema==2.16.2
filelock==3.8.0
fitz==0.0.1.dev2
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==23.1.21
fonttools==4.29.0
frozenlist==1.3.3
funcy==1.18
future==0.18.2
gast==0.4.0
gdown==4.7.1
gensim==3.8.3
ghostscript==0.7
gitdb==4.0.9
GitPython==3.1.30
google-api-core==2.11.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.6.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
h5py==3.8.0
htmlmin==0.1.12
httplib2==0.21.0
huggingface-hub==0.10.1
hyperlink==21.0.0
idna==3.3
ImageHash==4.3.0
imbalanced-learn==0.7.0
importlib-metadata==5.2.0
importlib-resources==5.10.2
incremental==22.10.0
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.6.1
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.17.3
jupyter_client==7.3.5
jupyter_core==5.1.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.0.2
keras==2.11.0
kiwisolver==1.3.2
kmodes==0.12.2
langcodes==3.3.0
Levenshtein==0.20.7
libclang==15.0.6.1
lightgbm==3.3.5
llvmlite==0.37.0
looseversion==1.0.3
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.0.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.1
mistune==0.8.4
mlflow==2.1.1
mlxtend==0.19.0
multidict==6.0.4
multimethod==1.8
murmurhash==1.0.8
mypy-extensions==0.4.3
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.4
networkx==2.8.6
nibabel==5.0.0
nipype==1.8.5
nltk==3.7
numba==0.54.1
numexpr==2.8.4
numpy==1.22.1
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
pandas-profiling==3.3.0
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.9.0
pathy==0.6.2
patsy==0.5.2
pdfminer.six==20221105
pdfplumber==0.7.6
phik==0.12.2
pickleshare==0.7.5
Pillow==9.0.0
pke-tool==1.8.1
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==2.4.1
plotly==5.13.0
preshed==3.0.7
prometheus-client==0.13.1
prompt-toolkit==3.0.26
Protego==0.2.1
proto-plus==1.22.2
protobuf==3.19.4
prov==2.0.0
psutil==5.9.2
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaret==2.3.10
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
pydeck==0.8.0
PyDispatcher==2.0.7
pydot==1.4.2
Pygments==2.11.2
PyJWT==2.6.0
pyLDAvis==3.2.2
Pympler==1.0.1
PyMuPDF==1.21.1
pynndescent==0.5.8
pyod==1.0.7
pyOpenSSL==23.0.0
pyparsing==3.0.7
PyPDF2==2.12.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyrsistent==0.18.1
PySocks==1.7.1
pyspark==3.3.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==303
pywinpty==2.0.1
pyxnat==1.5
PyYAML==5.4.1
pyzmq==22.3.0
querystring-parser==1.2.4
queuelib==1.6.2
rapidfuzz==2.11.1
rdflib==6.2.0
regex==2022.9.13
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.3.1
rsa==4.9
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
Scrapy==2.8.0
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
sense2vec==2.0.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
service-identity==21.1.0
shap==0.41.0
simplejson==3.18.1
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==2.3.9
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.46
sqlparse==0.4.3
srsly==1.0.6
stack-data==0.1.4
statsmodels==0.13.2
streamlit==1.18.1
tabula-py==2.6.0
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.13.1
testpath==0.5.0
textblob==0.17.1
thinc==7.4.6
threadpoolctl==3.1.0
tinycss2==1.2.1
tldextract==3.4.0
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.0
toolz==0.11.2
torch==1.12.1
torchvision==0.13.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
traits==6.3.2
transformers==4.23.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typer==0.4.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.2
umap-learn==0.5.3
Unidecode==1.3.6
urllib3==1.26.8
validators==0.18.2
virtualenv==20.16.3
visions==0.7.5
w3lib==2.1.1
waitress==2.1.2
Wand==0.6.11
wasabi==0.10.1
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wordcloud==1.8.2.2
wrapt==1.14.1
yarl==1.8.2
yellowbrick==1.3.post1
zipp==3.11.0
zope.interface==6.0
